backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
whitenoise>=6.6
Pillow>=10.0
reportlab>=4.0
python-dateutil>=2.8
